Found a cool maze that is triciking all my search algos
This maze currently is tricking all of our existing search
algorithms, as they prefer to move straight to the goal without
taking the teleporter right next to it. maze_5.txt is proof
that theseus.rb can handle teleportation, so it's the heuristics
that needs changing. And I'm not sure how to do that? After all,
the existing heurisitics were just given to me on a sliver platter...
but now I need to invent my own (or search for one heurisitc already
pre-made).